# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ke Klinik REST ful API Documentation",
        "description": "Description endpoint rest full api project ke klinik",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/service": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Add new Service",
                "description": "Endpoint to add new service with name",
                "operationId": "addService",
                "requestBody": {
                    "description": "Request addService needed field name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "validation field name: required, string, unique",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service add successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service add successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "9345"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Kontrol"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-09 07:59:39"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-09 07:59:39"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message field error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Route not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
...
}